refactor(web): migrate Avatar and Field onto @e2a/ui
CounterpartyAvatar was a near-duplicate of @e2a/ui Avatar (same FNV-ish
hash, --av palette, and initials); replace its six usages and delete the
file. The labeled-input Field was also duplicated — swap its four call
sites to @e2a/ui Field and trim Field.tsx down to the domain-specific
DNSRecord/DNSField helpers (renamed DnsRecord.tsx).

Minor deliberate differences: @e2a/ui Field focus ring uses the amber
--accent/--focus-ring token (vs the old green Tailwind accent) and
--r-lg radius; label renders at 13px vs 14px.

Verified: tsc --noEmit clean, eslint clean, 874 web tests pass, and the
full next build (static export) succeeds.
