+++ web/src/app/(app)/get-started/_components/DomainSelector.tsx
@@ -1,7 +1,7 @@
 "use client";
 
 import { useState } from "react";
-import { Field } from "../../../components/Field";
+import { Field } from "@e2a/ui";
 import { registerDomain } from "../../../components/onboarding/api";
 import { canReceive, isValidDomain } from "../../../components/onboarding/state";
 import { track } from "../../../components/onboarding/analytics";
