+++ web/src/app/components/DnsRecord.tsx
@@ -2,33 +2,6 @@
 
 import { useState } from "react";
 
-export function Field({
-  label,
-  hint,
-  ...props
-}: {
-  label: string;
-  hint?: React.ReactNode;
-  placeholder: string;
-  value: string;
-  onChange: (v: string) => void;
-  type?: string;
-}) {
-  return (
-    <div>
-      <label className="block text-sm font-medium mb-1.5">{label}</label>
-      <input
-        type={props.type || "text"}
-        placeholder={props.placeholder}
-        value={props.value}
-        onChange={(e) => props.onChange(e.target.value)}
-        className="w-full px-3 py-2.5 border border-border rounded-lg text-sm bg-surface focus:outline-none focus:ring-2 focus:ring-accent/30 focus:border-accent transition"
-      />
-      {hint && <p className="mt-1 text-xs text-muted">{hint}</p>}
-    </div>
-  );
-}
-
 export function DNSRecord({
   type,
   label,
